requirements.txt
# Django framework
Django>=4.2,<5.0

# Django REST Framework
djangorestframework>=3.15,<4.0

# Kafka client cho Python
kafka-python>=2.1,<3.0

# Celery cho async task processing
celery>=5.3,<6.0
redis>=5.0,<6.0

# Faker để tạo dữ liệu giả
Faker>=23.0

# HTTP requests
requests>=2.31

# CORS headers
django-cors-headers>=4.0

# Environment variables
python-dotenv>=1.1

# Production server
gunicorn>=21.2

# Logging
python-json-logger>=2.0

# Retry mechanism
tenacity>=8.2